Add Spring Security

+++ src/main/java/com/calendarEvents/CalendarImpEvents/Controllers/NavbarController.java
@@ -1,40 +1,50 @@
 package com.calendarEvents.CalendarImpEvents.Controllers;
 
+import com.calendarEvents.CalendarImpEvents.models.Enams.Role;
 import com.calendarEvents.CalendarImpEvents.models.Events;
+import com.calendarEvents.CalendarImpEvents.models.User;
 import com.calendarEvents.CalendarImpEvents.repositories.EventsRepositiry;
-import jdk.jfr.Event;
+import com.calendarEvents.CalendarImpEvents.repositories.UserRepository;
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.propertyeditors.CustomDateEditor;
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
-import org.springframework.web.bind.WebDataBinder;
 import org.springframework.web.bind.annotation.*;
 
-import java.text.SimpleDateFormat;
+import java.security.Principal;
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
 import java.util.ArrayList;
-import java.util.Date;
 import java.util.List;
-import java.util.Optional;
 
 @Controller
+@Slf4j
+@RequiredArgsConstructor
 public class NavbarController {
 
     @Autowired
     private EventsRepositiry eventsRepository;
+    @Autowired
+    private UserRepository userRepository;
+
 
     @GetMapping("/events")
-    public String events(Model model) {
-        Iterable<Events> events = eventsRepository.findAll();
+    public String events(Model model, Principal principal) {
+        User currentUser = userRepository.findByUsername(principal.getName()).orElseThrow(() -> new RuntimeException("User not found"));
+        Iterable<Events> events;
+        if (currentUser.hasRole(String.valueOf(Role.ROLE_ADMIN))) {
+            events = eventsRepository.findAll();
+        } else {
+            events = eventsRepository.findEventsByOwner(currentUser);
+        }
         model.addAttribute("events", events);
         return "events";
     }
 
     @GetMapping("/events/{date}")
     public String getEventsByDate(@PathVariable String date, Model model) {
-//        DateTimeFormatter dateTimeFormatter = DateTimeFormatter.ofPattern("YYYY-MM-DD");
-//        LocalDate localDate = LocalDate.parse(date, dateTimeFormatter);
         LocalDate localDate = LocalDate.parse(date, DateTimeFormatter.ISO_LOCAL_DATE);
         List<Events> event = eventsRepository.findByDate(localDate);
         model.addAttribute("localDate", localDate);
@@ -43,44 +53,39 @@ public class NavbarController {
     }
 
     @PostMapping("/events/{date}")
-    public String eventsAdd(@PathVariable String date, @RequestParam String title,@RequestParam String description, Model model) {
+    public String eventsAdd(@PathVariable String date, @RequestParam String title, @RequestParam String description, Model model, Principal principal) {
+        User currentUser = userRepository.findByUsername(principal.getName()).orElseThrow(() -> new RuntimeException("User not found"));
         LocalDate localDate = LocalDate.parse(date, DateTimeFormatter.ISO_LOCAL_DATE);
         Events event = new Events(title, localDate, description);
+        event.setOwner(currentUser);
         eventsRepository.save(event);
-//        boolean matches = date.matches("\\d{4}-\\d{2}-\\d{2}");
-//        LocalDate localDate = LocalDate.parse(date, DateTimeFormatter.ISO_LOCAL_DATE);
-//        Events event = new Events(title, localDate, description);
-//        eventsRepository.save(event);
         return "redirect:/events";
     }
 
     @GetMapping("/event/{id}")
-    public String eventsDetails(@PathVariable Long id, Model model) {
-        if (!eventsRepository.existsById(id)){
-            return "redirect:/events";
-        }
-        Optional<Events> events = eventsRepository.findById(id);
-        ArrayList<Events> result = new ArrayList<>();
-        events.ifPresent(result::add);
-        model.addAttribute("events", result);
+    public String eventsDetails(@PathVariable Long id, Model model, Principal principal) {
+        Events event = eventsRepository.findById(id).orElseThrow();
+        User currentUser = userRepository.findByUsername(principal.getName())
+                .orElseThrow(() -> new RuntimeException("User not found"));
+        checkUserAccess(event, currentUser);
+        model.addAttribute("events", event);
         return "events-details";
     }
 
     @GetMapping("/events/{id}/edit")
-    public String eventsEdit(@PathVariable Long id, Model model) {
-        if (!eventsRepository.existsById(id)){
-            return "redirect:/events";
-        }
-        Optional<Events> events = eventsRepository.findById(id);
-        ArrayList<Events> result = new ArrayList<>();
-        events.ifPresent(result::add);
-        model.addAttribute("events", result);
+    public String eventsEdit(@PathVariable Long id, Model model, Principal principal) {
+        Events event = eventsRepository.findById(id).orElseThrow();
+        User currentUser = userRepository.findByUsername(principal.getName()).orElseThrow(() -> new RuntimeException("User not found"));
+        checkUserAccess(event, currentUser);
+        model.addAttribute("events", event);
         return "events-edit";
     }
 
     @PostMapping("/events/{id}/edit")
-    public String eventsUpdate(@PathVariable Long id, @RequestParam String title,@RequestParam String description, Model model) {
+    public String eventsUpdate(@PathVariable Long id, @RequestParam String title,@RequestParam String description, Model model, Principal principal) {
+        User currentUser = userRepository.findByUsername(principal.getName()).orElseThrow(() -> new RuntimeException("User not found"));
         Events event = eventsRepository.findById(id).orElseThrow(); // .orElseThrow() исключение если запись не найдена
+        checkUserAccess(event, currentUser);
         event.setTitle(title);
         event.setDescription(description);
         eventsRepository.save(event);
@@ -88,8 +93,10 @@ public class NavbarController {
     }
 
     @PostMapping("/events/{id}/remove")
-    public String eventsDelete(@PathVariable Long id, Model model) {
+    public String eventsDelete(@PathVariable Long id, Model model, Principal principal) {
+        User currentUser = userRepository.findByUsername(principal.getName()).orElseThrow(() -> new RuntimeException("User not found"));
         Events event = eventsRepository.findById(id).orElseThrow(); // .orElseThrow() исключение если запись не найдена
+        checkUserAccess(event, currentUser);
         eventsRepository.delete(event);
         return "redirect:/events";
     }
@@ -107,6 +114,12 @@ public class NavbarController {
     public String contacts(Model model) {
         model.addAttribute("title", "Связь с нами");
         return "contacts";
+    }
+
+    private void checkUserAccess(Events event, User currentUser) {
+        if (!currentUser.equals(event.getOwner()) && !currentUser.hasRole(Role.ROLE_ADMIN.name())) {
+            throw new AccessDeniedException("Access denied");
+        }
     }
 
 }
